Add collision End

+++ Assets/Components/Position.cs
@@ -2,8 +2,8 @@
 
 public class Position : MonoBehaviour {
 	// Advice: FYFY component aims to contain only public members (according to Entity-Component-System paradigm).
-	public int x;
-	public int y;
-	public int targetX = -1;
-	public int targetY = -1;
+	public float x;
+	public float y;
+	public float targetX = -1;
+	public float targetY = -1;
 }
